Module secp256k1zkp::aggsig

Aggregated Signature (a.k.a. Schnorr) Functionality

Structs

AggSigContext

Manages an instance of an aggsig multisig context, and provides all methods to act on that context

Constants

ZERO_256

The 256 bits 0

Functions

add_signatures_single

Single-Signer addition of Signatures Returns: Ok(Signature) on success In: sig1: sig1 to add sig2: sig2 to add pubnonce_total: sum of public nonces

export_secnonce_single

Single-Signer (plain old Schnorr, sans-multisig) export nonce Returns: Ok(SecretKey) on success In: msg: the message to sign seckey: the secret key

sign_single

Single-Signer (plain old Schnorr, sans-multisig) signature creation Returns: Ok(Signature) on success In: msg: the message to sign seckey: the secret key extra: if Some(), add this key to s secnonce: if Some(SecretKey), the secret nonce to use. If None, generate a nonce pubnonce: if Some(PublicKey), overrides the public nonce to encode as part of e final_nonce_sum: if Some(PublicKey), overrides the public nonce to encode as part of e

verify_batch

Batch Schnorr signature verification Returns: true on success In: sigs: The signatures msg: The messages to verify pubkey: The public keys

verify_single

Single-Signer (plain old Schnorr, sans-multisig) signature verification Returns: Ok(Signature) on success In: sig: The signature msg: the message to verify pubnonce: if Some(PublicKey) overrides the public nonce used to calculate e pubkey: the public key pubkey_total: The total of all public keys (for the message in e) is_partial: whether this is a partial sig, or a fully-combined sig