Function secp256k1zkp::aggsig::sign_single

pub fn sign_single(
    secp: &Secp256k1, 
    msg: &Message, 
    seckey: &SecretKey, 
    secnonce: Option<&SecretKey>, 
    extra: Option<&SecretKey>, 
    pubnonce: Option<&PublicKey>, 
    pubkey_for_e: Option<&PublicKey>, 
    final_nonce_sum: Option<&PublicKey>
) -> Result<Signature, Error>

Single-Signer (plain old Schnorr, sans-multisig) signature creation Returns: Ok(Signature) on success In: msg: the message to sign seckey: the secret key extra: if Some(), add this key to s secnonce: if Some(SecretKey), the secret nonce to use. If None, generate a nonce pubnonce: if Some(PublicKey), overrides the public nonce to encode as part of e final_nonce_sum: if Some(PublicKey), overrides the public nonce to encode as part of e