Function secp256k1zkp::aggsig::verify_single

pub fn verify_single(
    secp: &Secp256k1, 
    sig: &Signature, 
    msg: &Message, 
    pubnonce: Option<&PublicKey>, 
    pubkey: &PublicKey, 
    pubkey_total_for_e: Option<&PublicKey>, 
    extra_pubkey: Option<&PublicKey>, 
    is_partial: bool
) -> bool

Single-Signer (plain old Schnorr, sans-multisig) signature verification Returns: Ok(Signature) on success In: sig: The signature msg: the message to verify pubnonce: if Some(PublicKey) overrides the public nonce used to calculate e pubkey: the public key pubkey_total: The total of all public keys (for the message in e) is_partial: whether this is a partial sig, or a fully-combined sig